move legacy code into a folder and make simulation parameters class.
