+++ pysac/io/legacy/util.py
